test: split ACP manager runtime config coverage
